Custom ball bounciness
A sphere was created in the (named as BouncyObj) which has a trigger circle collider and is a dynamic rigibody.
To avoid depending on physics materials and simulations, sensors around the BouncyObj are being used to implement a custom bouncy effect.
One thats effects are in complete control of the developers.