Fixed infinite side bounce problem and then fixed jump problem that followed
Since the ground isnt 100% plane, but the y velocity of the bouncy object becomes static when grounded, there are time a OnTrigger event is fired by accident, which affects the overall bounce of the object.
The solution so far is tweaked to fit with the level, but a better solution should be implemented for future development.